feat: arenaCreationManager class to handle all arena creation utilities
