+++ src/Frontend/Pages/LobbyConfigPage.tsx
@@ -1,10 +1,5 @@
-import { EthConnection } from '@darkforest_eth/network';
-import { EthAddress } from '@darkforest_eth/types';
 import React, { useEffect, useMemo, useReducer, useState } from 'react';
 import { Route, Switch, useHistory } from 'react-router-dom';
-import { ContractsAPI } from '../../Backend/GameLogic/ContractsAPI';
-import { createAndInitArena } from '../../Backend/Utils/Arena';
-import { LobbyAdminTools } from '../../Backend/Utils/LobbyAdminTools';
 import { MinimapConfig } from '../Panes/Lobbies/MinimapUtils';
 import {
   InvalidConfigError,
@@ -17,58 +12,47 @@ import { LobbyMapSelectPage } from './LobbyMapSelectPage';
 import { LobbyWorldSettingsPage } from './LobbyWorldSettingsPage';
 import { LobbyConfirmPage } from './LobbyConfirmPage';
 import { LobbyMapEditor } from './LobbyMapEditor';
-import { getAllTwitters } from '../../Backend/Network/UtilityServerAPI';
-import { DEFAULT_PLANET } from '../Panes/Lobbies/LobbiesUtils';
-import { Toast } from '../Components/Toast';
 import _ from 'lodash';
+import { ArenaCreationManager } from '../../Backend/GameLogic/ArenaCreationManager';
+import { Toast } from '../Components/Toast';
 
 type Status = 'waitingForCreate' | 'creating' | 'created' | 'errored' | undefined;
 
 const BULK_CREATE_CHUNK_SIZE = 5;
 
 export function LobbyConfigPage({
-  contractsAPI,
-  connection,
-  ownerAddress,
+  arenaCreationManager,
   startingConfig,
   root,
 }: {
-  contractsAPI: ContractsAPI;
-  connection: EthConnection;
-  ownerAddress: EthAddress;
+  arenaCreationManager: ArenaCreationManager;
   startingConfig: LobbyInitializers;
   root: string;
 }) {
   const [config, updateConfig] = useReducer(lobbyConfigReducer, startingConfig, lobbyConfigInit);
   const [minimapConfig, setMinimapConfig] = useState<MinimapConfig | undefined>();
-  const [lobbyAdminTools, setLobbyAdminTools] = useState<LobbyAdminTools>();
   const [lobbyTx, setLobbyTx] = useState<string | undefined>();
   const [status, setStatus] = useState<Status>(undefined);
   const [error, setError] = useState<string | undefined>(undefined);
   const [playerTwitter, setPlayerTwitter] = useState<string | undefined>();
 
   const createDisabled = status === 'creating' || status === 'created';
-  const creating = status === 'creating' || (status === 'created' && !lobbyAdminTools?.address);
-  const created = status === 'created' && lobbyAdminTools?.address;
+  const creating =
+    status === 'creating' || (status === 'created' && !arenaCreationManager?.address);
+  const created = status === 'created' && arenaCreationManager?.address;
 
+  // once admin tools are created, create and reveal
   useEffect(() => {
     async function doCreateReveal() {
       await bulkCreateAndRevealPlanets();
       setStatus('created');
     }
-    if (lobbyAdminTools && !created) {
+    if (arenaCreationManager && !created) {
       doCreateReveal();
     }
-  }, [lobbyAdminTools]);
+  }, [arenaCreationManager]);
 
-  useEffect(() => {
-    async function fetchTwitters() {
-      const allTwitters = await getAllTwitters();
-      setPlayerTwitter(allTwitters[ownerAddress]);
-    }
-    fetchTwitters();
-  }, []);
-
+  // set error when theres an admin planets warning
   useEffect(() => {
     if (config.ADMIN_PLANETS.warning) {
       setError(config.ADMIN_PLANETS.warning);
@@ -76,21 +60,17 @@ export function LobbyConfigPage({
   }, [config.ADMIN_PLANETS.warning]);
 
   async function bulkCreateAndRevealPlanets() {
-    if (!lobbyAdminTools) {
-      setError("You haven't created a lobby.");
-      throw new Error('No lobby');
-    }
     if (!config.ADMIN_PLANETS.currentValue) {
       setError('no planets staged');
       throw new Error('No planets staged');
     }
     let planets = config.ADMIN_PLANETS.currentValue;
-
+    const initializers = toInitializers(config);
     let i = 0;
     while (i < planets.length) {
       try {
         const chunk = planets.slice(i, i + BULK_CREATE_CHUNK_SIZE);
-        await lobbyAdminTools.bulkCreateAndReveal(chunk, toInitializers(config));
+        await arenaCreationManager.bulkCreateAndReveal(chunk, initializers);
         updateConfig({
           type: 'ADMIN_PLANETS',
           value: undefined,
@@ -103,22 +83,20 @@ export function LobbyConfigPage({
         console.error('Error creating and revealing planets:', err);
         if (err instanceof InvalidConfigError) {
           setError(`Invalid ${err.key} value ${err.value ?? ''} - ${err.message}`);
-          throw new Error(`Invalid ${err.key} value ${err.value ?? ''} - ${err.message}`)
+          throw new Error(`Invalid ${err.key} value ${err.value ?? ''} - ${err.message}`);
         } else {
           setError(err?.message || 'Something went wrong. Check your dev console.');
           throw new Error(err?.message || 'Something went wrong. Check your dev console.');
-
         }
       }
     }
     setStatus('created');
   }
-
   async function validateAndCreateLobby() {
     try {
       setStatus('creating');
       const initializers = toInitializers(config);
-      await createLobby(initializers);
+      await arenaCreationManager.createAndInitArena(initializers);
     } catch (err) {
       setStatus('errored');
       console.error(err);
@@ -130,33 +108,11 @@ export function LobbyConfigPage({
     }
   }
 
-  async function createLobby(config: LobbyInitializers) {
-    try {
-      const { owner, lobby, startTx } = await createAndInitArena({
-        config,
-        contractsAPI,
-        ethConnection: connection,
-      });
-
-      setLobbyTx(startTx?.hash);
-
-      if (owner === ownerAddress) {
-        if (!connection) {
-          throw 'error: no connection';
-        }
-        const lobbyAdminTools = await LobbyAdminTools.create(lobby, connection);
-        setLobbyAdminTools(lobbyAdminTools);
-      }
-    } catch (e) {
-      console.log(e);
-      throw new Error(e?.message || `Failed to create lobby.`);
-    }
-  }
-
   const onMapChange = useMemo(() => {
     return _.debounce((config: MinimapConfig) => setMinimapConfig(config), 500);
   }, [setMinimapConfig]);
 
+  // update entire map if something in map changes
   useEffect(() => {
     onMapChange({
       worldRadius: config.WORLD_RADIUS_MIN.currentValue,
@@ -168,7 +124,7 @@ export function LobbyConfigPage({
       perlinThreshold2: config.PERLIN_THRESHOLD_2.currentValue,
       perlinThreshold3: config.PERLIN_THRESHOLD_3.currentValue,
       stagedPlanets: config.ADMIN_PLANETS.currentValue || [],
-      createdPlanets: lobbyAdminTools?.planets || [],
+      createdPlanets: arenaCreationManager?.planets || [],
       dot: 4,
     });
   }, [
@@ -182,7 +138,7 @@ export function LobbyConfigPage({
     config.PERLIN_THRESHOLD_2.currentValue,
     config.PERLIN_THRESHOLD_3.currentValue,
     config.ADMIN_PLANETS.currentValue,
-    lobbyAdminTools,
+    arenaCreationManager,
   ]);
 
   return (
@@ -198,10 +154,9 @@ export function LobbyConfigPage({
       <Switch>
         <Route path={root} exact={true}>
           <LobbyMapSelectPage
-            address={ownerAddress}
+            address={arenaCreationManager.address}
             startingConfig={startingConfig}
             updateConfig={updateConfig}
-            lobbyAdminTools={lobbyAdminTools}
             createDisabled={createDisabled}
             root={root}
             setError={setError}
@@ -209,13 +164,12 @@ export function LobbyConfigPage({
         </Route>
         <Route path={`${root}/confirm`}>
           <LobbyConfirmPage
-            lobbyAdminTools={lobbyAdminTools}
+            arenaCreationManager={arenaCreationManager}
             minimapConfig={minimapConfig}
             config={config}
             onUpdate={updateConfig}
             createDisabled={createDisabled}
             root={root}
-            ownerAddress={ownerAddress}
             lobbyTx={lobbyTx}
             onError={setError}
             created={created}
@@ -231,7 +185,7 @@ export function LobbyConfigPage({
             createDisabled={createDisabled}
             root={root}
             minimapConfig={minimapConfig}
-            lobbyAdminTools={lobbyAdminTools}
+            arenaCreationManager={arenaCreationManager}
           />
         </Route>
         <Route path={`${root}/edit-map`}>
@@ -241,9 +195,8 @@ export function LobbyConfigPage({
             createDisabled={createDisabled}
             root={root}
             minimapConfig={minimapConfig}
-            lobbyAdminTools={lobbyAdminTools}
+            arenaCreationManager={arenaCreationManager}
             onError={setError}
-            ownerAddress={ownerAddress}
           />
         </Route>
       </Switch>
