+++ src/Frontend/Pages/LobbyMapSelectPage.tsx
@@ -2,7 +2,6 @@ import { EthAddress, WorldCoords } from '@darkforest_eth/types';
 import React, { CSSProperties, useState } from 'react';
 import { useHistory } from 'react-router-dom';
 import styled from 'styled-components';
-import { LobbyAdminTools } from '../../Backend/Utils/LobbyAdminTools';
 import { Spacer } from '../Components/CoreUI';
 import { Minimap } from '../Components/Minimap';
 import { Toast } from '../Components/Toast';
@@ -16,11 +15,10 @@ export const LobbyMapSelectPage: React.FC<{
   address : EthAddress;
   startingConfig: LobbyInitializers;
   updateConfig: React.Dispatch<LobbyAction>;
-  lobbyAdminTools: LobbyAdminTools | undefined;
   createDisabled: boolean;
   root: string;
   setError: (error: string) => void;
-}> = ({ address, startingConfig, updateConfig, lobbyAdminTools, createDisabled, root, setError }) => {
+}> = ({ address, startingConfig, updateConfig, createDisabled, root, setError }) => {
   const mapSizePx = '250px';
   const history = useHistory();
 
@@ -40,7 +38,7 @@ export const LobbyMapSelectPage: React.FC<{
       perlinThreshold2: config.PERLIN_THRESHOLD_2,
       perlinThreshold3: config.PERLIN_THRESHOLD_3,
       stagedPlanets: config.ADMIN_PLANETS || [],
-      createdPlanets: lobbyAdminTools?.planets || [],
+      createdPlanets: [],
       dot: 10,
     } as MinimapConfig;
   }
